# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CourseHub API",
        "description": "Complete API documentation for CourseHub platform",
        "contact": {
            "name": "CourseHub Support",
            "email": "admin@coursehub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/courses": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Retrieve a list of courses",
                "description": "Retrieve a paginated list of courses.",
                "operationId": "fb9d5ed8eebcf64545eccbec4e07e665",
                "parameters": [
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search courses by title or description.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort courses by field. Use \"-\" prefix for descending order.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "-title",
                                "price",
                                "-price",
                                "created_at",
                                "-created_at"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Relations to include.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "author,lessons_count"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course list retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "summary": "Create a course",
                "description": "Create a new course.",
                "operationId": "b923cafaec6b13f5d12afd86e8a55be9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated."
                    },
                    "403": {
                        "description": "User does not have permissions."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/courses/{course}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "summary": "Retrieve course details",
                "description": "Retrieve detailed information about a specific course.",
                "operationId": "ac9df8579adf9db70c70ca96b5800656",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "math-101"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Course not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "summary": "Remove a course",
                "description": "Remove the specified course.",
                "operationId": "b2ad98bf222fcb881591892f6ea118cc",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "math-101"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course deleted successfully."
                    },
                    "401": {
                        "description": "User is unauthenticated."
                    },
                    "403": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Course not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Course"
                ],
                "summary": "Update a course",
                "description": "Update the specified course.",
                "operationId": "f4c07044a4f22f76356c7c367d19f56d",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "math-101"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated."
                    },
                    "403": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Course not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateCourseRequest": {
                "title": "Create Course Request",
                "description": "Request payload for creating a new course.",
                "required": [
                    "title",
                    "type",
                    "price"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the course.",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Math 101"
                    },
                    "slug": {
                        "description": "Slug of the course. (optional, must be unique if provided)",
                        "type": "string",
                        "maxLength": 255,
                        "pattern": "^[a-z0-9-]+$",
                        "example": "math-101",
                        "nullable": true
                    },
                    "description": {
                        "description": "Description of the course.",
                        "type": "string",
                        "maxLength": 5000,
                        "example": "A basic mathematics course",
                        "nullable": true
                    },
                    "type": {
                        "description": "Type of the course.",
                        "type": "string",
                        "enum": [
                            "offline",
                            "online",
                            "video"
                        ],
                        "example": "online"
                    },
                    "price": {
                        "description": "Price of the course.",
                        "type": "string",
                        "format": "float",
                        "maximum": 99999999.99,
                        "minimum": 0,
                        "example": 199.99
                    }
                },
                "type": "object"
            },
...
            "UpdateCourseRequest": {
                "title": "Update Course Request",
                "description": "Request payload for updating the course.",
                "required": [],
                "properties": {
                    "title": {
                        "description": "Title of the course.",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Math 101"
                    },
                    "slug": {
                        "description": "Slug of the course. (must be unique)",
                        "type": "string",
                        "maxLength": 255,
                        "pattern": "^[a-z0-9-]+$",
                        "example": "math-101"
                    },
                    "description": {
                        "description": "Description of the course.",
                        "type": "string",
                        "example": "A basic mathematics course",
                        "nullable": true
                    },
                    "price": {
                        "description": "Price of the course.",
                        "type": "string",
                        "format": "float",
                        "maximum": 99999999.99,
                        "minimum": 0,
                        "example": 199.99
                    }
                },
                "type": "object"
            },
            "Author": {
                "title": "Author Schema",
                "description": "Details of a author returned by the API",
                "required": [
                    "id",
                    "name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "description": "The user ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The user name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "slug": {
                        "description": "The user slug",
                        "type": "string",
                        "example": "john-doe"
                    }
                },
                "type": "object"
            },
            "CourseResponse": {
                "title": "Course Response",
                "description": "Data of a specific course.",
                "required": [
                    "id",
                    "author_id",
                    "title",
                    "slug",
                    "description",
                    "type",
                    "price",
                    "image_url",
                    "is_published",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the course",
                        "type": "integer",
                        "example": 1
                    },
                    "author_id": {
                        "description": "Author unique identifier of the course",
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Author"
                            }
                        ],
                        "nullable": true,
                        "description": "Author of the course"
                    },
                    "title": {
                        "description": "Title of the course",
                        "type": "string",
                        "example": "Math 101"
                    },
                    "slug": {
                        "description": "Slug of the course",
                        "type": "string",
                        "example": "math-101"
                    },
                    "description": {
                        "description": "Description of the course",
                        "type": "string",
                        "example": "A basic mathematics course",
                        "nullable": true
                    },
                    "type": {
                        "description": "Type of the course",
                        "type": "string",
                        "enum": [
                            "online",
                            "offline",
                            "video"
                        ],
                        "example": "online"
                    },
                    "price": {
                        "description": "Price of the course",
                        "type": "string",
                        "format": "float",
                        "example": 99.99
                    },
                    "image_url": {
                        "description": "Image url of the course",
                        "type": "string",
                        "format": "url",
                        "example": "http://loclhost:8080/storage/courses/course1.png",
                        "nullable": true
                    },
                    "is_published": {
                        "description": "Status of the course",
                        "type": "boolean",
                        "example": true
                    },
                    "lessons_count": {
                        "description": "Total number of course lessons",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation date of the course",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-20T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Modification date of the course",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-25T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}